+++ frontend/src/App.css
@@ -3,29 +3,34 @@
 }
 
 html {
-  border: 1.5rem solid lightgreen;
+  border: 3vh solid lightgreen;
   height: 100vh;
+  position: relative;
+  display:flex;
+  flex-direction: column;
 }
 
-.App {
+.center {
   text-align: center;
 }
 
-.logo {
+/* Login Page */
+
+.loginLogo {
   margin-top: 4rem;
   width: 80%;
 }
 
-.header {
+.loginHeader {
   margin-top: 5rem;
 }
 
-.inputContainer {
+.loginInputContainer {
   display: flex;
   justify-content: center;
 }
 
-.inputField {
+.loginInputField {
   text-align: center;
   width: 25%;
   padding: 1rem;
@@ -39,3 +44,90 @@ html {
   background-color: lightgreen;
   border: none;
 }
+
+/* Main Page */
+
+.mainPage {
+  display: flex;
+  flex-direction: column;
+  height: 94vh;
+}
+
+.header {
+  height: 10vh;
+  margin-top: 0px;
+  background-color: lightgray;
+}
+
+.logo {
+  width: 140px;
+}
+
+.switch {
+  min-width: 6rem; 
+  background-color: lightgreen;
+  border: none;
+}
+
+.footer {
+  background-color: lightgray;
+  height: 30vh;
+  flex: 1;
+  display:flex;
+  flex-direction: column;
+}
+
+.nowPlaying {
+  width: 100%;
+  height: 50%;
+  flex: 1;
+  background-color: lightsalmon;
+}
+
+.nav {
+  height: 50%;
+}
+
+
+/*  Queue Window */
+
+.window {
+  padding-top: 25px;
+  overflow: auto;
+  height: 60vh;
+  background-color: rgba(0, 0, 0, 0.7);
+}
+
+.track {
+  height: 100px;
+}
+
+.songImage {
+  float: left;
+  margin-left: 25px;
+}
+
+.artistName{
+  display: inline-block;
+}
+
+.songName{
+  display: inline-block;
+}
+
+
+.upvote {
+  float: right;
+  margin-right: 10%;
+  margin-top: 25px;
+}
+
+.voteNumber {
+  float: right;
+  margin-right: 10%;
+}
+
+/*  Search Window */
+
+
+
