+++ frontend/src/components/QueueWindow.js
@@ -34,21 +34,18 @@ class QueueWindow extends Component {
     if (this.state.tracks) {
     const annonser = [...this.state.tracks];
     queue = annonser.map(annons => (
-      <div key={annons.id}>
-      <p><strong>{annons.name}</strong></p>
-        <p><em>{annons.artists[0].name}</em></p>
-        <img src={annons.album.images[2].url} />
-        <br/>
-        <h3> Votes: {annons.votes} </h3>
-        <button onClick={ () =>  this.upvote(annons)}> Upvote </button>
-        <br/>
-        <button onClick={ () =>  this.removeFromQueue(annons)}> Remove from queue </button>
-        <hr></hr>
+      <div key={annons.id} className="track">
+      <p className="songName"><strong>{annons.name}</strong></p>
+        <button className="upvote"onClick={ () =>  this.upvote(annons)}> Upvote </button>
+        <p className="artistName"><em>{annons.artists[0].name}</em></p>
+        <img className="songImage" src={annons.album.images[2].url} />
+        <p className="voteNumber"> Votes: {annons.votes} </p>
+        {/* <button onClick={ () =>  this.removeFromQueue(annons)}> Remove from queue </button>  */}
       </div>
     ))
   } 
     return (
-      <div className="App">
+      <div className="window">
         <h2> Queue </h2>
         <hr></hr>
         { queue }
